Add schema testing
Introduce integration testing for the API service. These tests expect
a live database, so the easiest way to run them is to use docker:

```
npm run docker-test
```

These tests let developers verify that the GraphQL API behaves as
expected given known states in the database - I think this is the most
important kind of testing we can do at this level.