Add auth and context
The motivation here was to add context, and the implementation ended
up being an express-jwt and HttpOnly cookie based approach to
authentication.

Authentication for GraphQL apps sucks in 2018. JWT feels like the
future, but the only way to securely transmit it is with an HttpOnly
cookie that is, by definition, not accessible in JavaScript and
therefore cannot be sent using bearer token authentication.

I'm eschewing Passport.js here because I think it's overkill for a
GraphQL service and I want to build something that makes more sense.